Moving PayButton handler to public controller
