+++ BTCPayServer/Views/Public/PayButtonTest.cshtml
@@ -3,7 +3,7 @@
     <div class="container" id="payButtonCtrl">
         <div class="row">
 
-            <form method="POST" action="http://127.0.0.1:14142/stores/4b1AnMtMNuxwG7io2dUcB3TrT3yUbgajjGfJjVfdhBDu/pay">
+            <form method="POST" action="http://127.0.0.1:14142/pay/4b1AnMtMNuxwG7io2dUcB3TrT3yUbgajjGfJjVfdhBDu">
                 <input type="hidden" name="price" value="10" />
                 <input type="hidden" name="currency" value="USD" />
                 <input type="image" src="http://127.0.0.1:14142/img/paybutton/pay.png" name="submit" style="width:209px" alt="Pay with BtcPay, Self-Hosted Bitcoin Payment Processor">
